bLIP-04: update from endorsement to accountable signal
Removed the `experiment_start/end` dates. This is done
because in the previous iteration with endorsement
signaling, the recommendation was for the sender to
set it to 1 and that could have had privacy concerns
when first deploying given that the default was to
downgrade the signal to 0. Now the recommended default
for both sending and forwarding nodes is to set
`accountable` to 0. Hence, the dates have been
removed as nodes can start setting the signal
without leaking information about being the sender.